+++ test/tests/UnusualTypeApplications.hs
@@ -1,0 +1,30 @@
+-- SYNTAX TEST "source.haskell" "Unusual type applications"
+
+
+f =  g @A
+--     ^^ meta.type-application.haskell
+f = [g]@A
+--     ^^ meta.type-application.haskell
+f = {g}@A
+--     ^^ meta.type-application.haskell
+f =  g,@A
+--     ^^ meta.type-application.haskell
+f =  g;@A
+--     ^^ meta.type-application.haskell
+f = g @(!A)
+--    ^^^^^ meta.type-application.haskell
+f = g @[!A]
+--    ^^^^^ meta.type-application.haskell
+
+f =   g@A
+--     ^^ - meta.type-application.haskell
+f =   +@A
+--     ^^ - meta.type-application.haskell
+f =  g@@A
+--     ^^ - meta.type-application.haskell
+f = g @@A
+--     ^^ - meta.type-application.haskell
+f = g @!A
+--     ^^ - meta.type-application.haskell
+f = (g)@A
+--     ^^ - meta.type-application.haskell