+++ test/tickets/T0073.hs
@@ -1,26 +1,26 @@
 -- SYNTAX TEST "source.haskell" "Visible type application"
 
 function = 
-    let l = [] @[Int] in
+    let l = [] @[( Int, f [ Bool ] )] in
 --             ^ meta.type-application.haskell
---               ^^^ storage.type.haskell 
+--                 ^^^ storage.type.haskell 
+--                      ^ variable.other.generic-type.haskell
     let x = "Hello" @String in
 --                  ^ meta.type-application.haskell 
 --                   ^^^^^^ storage.type.haskell
     let c@(Some pattern) = undefined in
---       ^ keyword.operator.haskell
 --       ^ - meta.type-application.haskell
 --         ^^^^ constant.other.haskell
-    let g = f @(g -> String) in
+    let g = f @( g ([]) -> String ) in
 --            ^ meta.type-application.haskell
---              ^ variable.other.generic-type.haskell
-    let h = k @_ @_ @Int
---            ^  ^ meta.type-application.haskell
+--               ^ variable.other.generic-type.haskell
+    let h = k @_ @a @Int
+--            ^  ^  ^ meta.type-application.haskell
+    let a = b @ c
+--            ^ - meta.type-application.haskell
     undefined
 
-
--- Visible kind application
-
-type family TF (a :: k) ( b :: l ) :: l where
-  TF @_ @(l :: Nat) A = B
---   ^  ^ meta.type-application.haskell
+g = f @Bool(A)
+--    ^ meta.type-application.haskell
+--     ^^^^ storage.type.haskell
+--          ^ - storage.type.haskell