+++ test/tickets/T0073b.hs
@@ -1,0 +1,4 @@
+-- SYNTAX TEST "source.haskell" "Visible kind application"
+
+type TF l = F @_ @(l :: Nat) A
+--            ^  ^ meta.type-application.haskell