+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - Address regression in LiquidHaskell highlighting ([#131](https://github.com/JustusAdam/language-haskell/issues/136)).
 - Fixed several issues involving multi-line type/data declarations,
   including allowing intervening comments ([#147](https://github.com/JustusAdam/language-haskell/pull/147)).
+- Add support for visible kind application (and miscellaneous fixes for visible type application).
 
 ## 3.0.0 - 26.04.2020